refactor(tag): make check-tag a separate directory
