+++ src/check-tag/check-tag.json
@@ -1,6 +1,6 @@
 {
   "component": true,
   "usingComponents": {
-    "t-icon": ".././icon/icon"
+    "t-icon": "../icon/icon"
   }
 }
