+++ src/check-tag/props.ts
