+++ src/check-tag/type.ts
@@ -6,14 +6,36 @@
 
 import { SizeEnum } from '../common/common';
 
-export interface TdTagProps {
+export interface TdCheckTagProps {
+  /**
+   * 标签选中的状态，默认风格（theme=default）才有选中态
+   */
+  checked?: {
+    type: BooleanConstructor;
+    value?: boolean;
+  };
+  /**
+   * 标签选中的状态，默认风格（theme=default）才有选中态，非受控属性
+   */
+  defaultChecked?: {
+    type: BooleanConstructor;
+    value?: boolean;
+  };
   /**
    * 标签是否可关闭
    * @default false
+   * @deprecated
    */
   closable?: {
     type: BooleanConstructor;
     value?: boolean;
+  };
+  /**
+   * 组件子元素；传入数组时：[选中内容，非选中内容]
+   */
+  content?: {
+    type: null;
+    value?: string | number | string[];
   };
   /**
    * 自定义组件样式
@@ -39,22 +61,16 @@ export interface TdTagProps {
     value?: ['t-class'];
   };
   /**
-   * 标签中的图标，可自定义图标呈现
+   * 标签图标
    */
   icon?: {
     type: null;
     value?: string | object;
   };
   /**
-   * 标签最大宽度，宽度超出后会出现省略号。示例：'50px' / 80
-   */
-  maxWidth?: {
-    type: null;
-    value?: string | number;
-  };
-  /**
    * 标签类型，有三种：方形、圆角方形、标记型
    * @default square
+   * @deprecated
    */
   shape?: {
     type: StringConstructor;
@@ -67,14 +83,6 @@ export interface TdTagProps {
   size?: {
     type: StringConstructor;
     value?: SizeEnum;
-  };
-  /**
-   * 组件风格，用于描述组件不同的应用场景
-   * @default default
-   */
-  theme?: {
-    type: StringConstructor;
-    value?: 'default' | 'primary' | 'warning' | 'danger' | 'success';
   };
   /**
    * 标签风格变体
