+++ src/tag/README.md
@@ -13,7 +13,7 @@ isComponent: true
 ```json
 "usingComponents": {
   "t-tag": "tdesign-miniprogram/tag/tag",
-  "t-check-tag": "tdesign-miniprogram/tag/check-tag"
+  "t-check-tag": "tdesign-miniprogram/check-tag/check-tag"
 }
 ```
