+++ src/tag/_example/checkable/index.json
@@ -2,6 +2,6 @@
   "component": true,
   "usingComponents": {
     "t-tag": "tdesign-miniprogram/tag/tag",
-    "t-check-tag": "tdesign-miniprogram/tag/check-tag"
+    "t-check-tag": "tdesign-miniprogram/check-tag/check-tag"
   }
 }
