+++ src/tag/_example/tag.json
@@ -1,8 +1,6 @@
 {
   "navigationBarTitleText": "Tag",
   "usingComponents": {
-    "t-tag": "tdesign-miniprogram/tag/tag",
-    "t-check-tag": "tdesign-miniprogram/tag/check-tag",
     "theme": "./theme",
     "size": "./size",
     "type": "./type",
